Remove  prefix
