+++ lib/erblint-github/linters.rb
@@ -1,3 +1,3 @@
 # frozen_string_literal: true
 
-Dir[File.join(__dir__, "linters", "github/**/*.rb")].sort.each { |file| require file }
+Dir[File.join(__dir__, "linters", "**/*.rb")].sort.each { |file| require file }
